Fix Rust Ruuter DSL/SQL compatibility: assign forward-refs, type casts, array serialisation
- Split same-block forward-reference assign steps across all affected DSLs:
  GET/v1/classifiers/classifier.yml, PUT/v1/classifiers.yml,
  PUT/v1/classifiers/value.yml, POST/v1/classifiers/value.yml,
  GET/templates + TEMPLATES classifier changed-fields templates
- Fix list_user_groups.sql and get_user_group.sql: wrap ARRAY() as
  ARRAY_TO_JSON() so Rust Resql serialises TEXT[] as a JSON array, and
  cast :organisation_id params with ::INTEGER / ::BIGINT to avoid
  'operator does not exist: text = integer' errors
- Fix get_user_group_users.sql: cast :user_organisation_id::INTEGER
- Fix PUT/v1/user-groups/permissions.yml: remove forward-ref to
  added_codes/removed_codes by inlining [] literals in log_content

+++ DSL/Resql/ljvis/POST/user_group/get_user_group_users.sql
@@ -77,7 +77,7 @@ WHERE
         )
     AND
     (
-    :user_organisation_id = 0
+    :user_organisation_id::INTEGER = 0
         OR l.organisation_id = :user_organisation_id::BIGINT
     )
     AND (
